% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_orcid.R
\name{fetch_orcid}
\alias{fetch_orcid}
\title{Get Author Info}
\usage{
fetch_orcid(orcid)
}
\arguments{
\item{orcid}{character. A vector of ORCID numbers.}
}
\value{
a data frame containing in the lines the information for each of the
ORCID numbers provided.
}
\description{
Get and organize author's personal and employement information
obtained from ORCID.
}
\details{
The functions obtains and organizes personal and employment
information using the \strong{rorcid} functions 'orcid_person' and
'orcid_employments'. Please check the help of these functions for more
detail.
}
\examples{

\dontrun{
  orcids <- c("0000-0002-5537-5294", NA, "0000-0002-1048-0138", "not found")
  orcid_info(orcids)
}   
 
}
\author{
Renato A. F. de Lima
}
