% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_ror.R
\name{fetch_ror}
\alias{fetch_ror}
\title{Get Metadata of ROR}
\usage{
fetch_ror(ror, output = NULL)
}
\arguments{
\item{ror}{a character conatinig the ROR IDs}

\item{output}{the type of output desired: record metadata or relationships.}
}
\value{
a data frame with the fields available for each ROR record.
}
\description{
Get all fields for a given ROR record
}
\details{
This function queries the ROR API (https://ror.readme.io/docs/rest-api) for
retrieving organization record by ROR ID.
}
\examples{

\dontrun{

  
  rors <- "05x5km989"
  fetch_ror(rors, output = "metadata")
  
}   
  
}
\author{
Renato A. F. de Lima
}
