% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orcid.R
\name{get_orcid}
\alias{get_orcid}
\title{Get ORCID}
\usage{
get_orcid(
  x,
  first.name = "FirstName",
  last.name = "LastName",
  middle.name = "MiddleName",
  orcid.name = "ORCID",
  keywords = NULL,
  rows = 20,
  clean = FALSE
)
}
\arguments{
\item{x}{a data frame containing at least two columns: first name and last
name}

\item{first.name}{character. The name of the column containing the first
name. Defaults to "FirstName".}

\item{last.name}{character. The name of the column containing the last name.
Defaults to "LastName".}

\item{middle.name}{character. The name of the column containing the middle
name (optional). Defaults to "MiddleName"}

\item{orcid.name}{character. The name of the column containing the ORCID
number (optional). Defaults to "ORCID".}

\item{keywords}{character. one or more keywords to filter the search output,
in the case of two or more ORCIDs}

\item{rows}{integer. The number of ORCID records to return, to be passed to
function rorcid::orcid_search. Defaults to 20.}

\item{clean}{logical. Should the function return only the exact name search?
Defaults to FALSE.}
}
\value{
the same input data frame with the missing (i.e. NA) ORCID, if found.
}
\description{
Get ORCID based on people's names.
}
\details{
If only one researcher is found the function returns a sole ORCID. If
there are two or more ORCID found, all ORCID are returned separated by a pipe
(i.e. "|").

The arguments 'keywords' can be used to narrow down the search if multiple
IDs are found. But it will only work if the reasearchers have included
keywords in their ORCID records, which is not always the case (see examples).

In addition, if argument 'clean' is TRUE, it will exclude those IDs which are
not exact matches, which can be useful to remove spurious hits. Also, you
may want to increase the number of results returned by each query if the names
provided are too common, to make sure that the name you are searching is
indeed returned in the query.

The package \strong{rorcid} which is used internally, requires an ORCID token or
that the user starts a ORCID session and authorise the package, so it can
query their API. So, make sure you have one of the options ready before using
the function.
}
\examples{

\dontrun{
  # Single ORCID record with the name
  df <- data.frame(FirstName = "Nicolas", LastName = "Casajus")
  get_orcid(df)
  
  # Multiple ORCID entries
  df <- data.frame(FirstName = "Renato", LastName = "Lima")
  get_orcid(df)
  get_orcid(df, clean = TRUE)
  get_orcid(df, keywords = "ecology")
}   
  
}
\author{
Renato A. F. de Lima
}
