% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_ror.R
\name{search_ror}
\alias{search_ror}
\title{Get ROR}
\usage{
search_ror(
  x,
  org.name = "OrganizationName",
  country.name = "CountryName",
  type = NULL,
  matching.type = NULL
)
}
\arguments{
\item{x}{a data frame containing at least one column: the organization name}

\item{org.name}{character. The name of the column containing the organization
name. Defaults to "OrganizationName".}

\item{country.name}{character. The name of the column containing the name of the organization country.
Defaults to "CountryName".}

\item{type}{character. The type of organizations for matching (see details)}

\item{matching.type}{character. The type of match performed (see details)}
}
\value{
the same input data frame with the missing (i.e. NA) ORCID, and
optionally the corresponding names, if found.
}
\description{
Get ROR ID based on the organisation name.
}
\details{
This function queries the ROR API (https://ror.readme.io/docs/rest-api)

If not NULL, the argument \code{type} can be one of the following: Education,
Healthcare, Company, Archive, Nonprofit, Government, Facility, and Other.

If not NULL, the argument \code{matching.type} can be one of the following:
\itemize{
\item PHRASE: the entire phrase matched to a variant of the organization's name
\item COMMON TERMS: the matching was done by comparing the words separately
\item FUZZY: the matching was done by fuzzy-comparing the words separately
\item HEURISTICS: "University of X" was matched to "X University"
\item ACRONYM: matching by acronym
}
}
\examples{

\dontrun{

  df <- data.frame(org.name = "USP", country.name = "Brazil")
  search_ror(df)
  
}   
  
}
\author{
Renato A. F. de Lima
}
